+++ src/core/application/ports/pagamento_port.rs
@@ -12,7 +12,7 @@ pub trait PagamentoPort: Send + Sync {
     fn processa_pagamento(
         &self,
         pedido_id: usize,
-        valor_pagamento: f32,
+        valor_pagamento: f64,
     ) -> Result<StatusPagamento, DomainError>;
 
     fn pagamento_status(
